Added simple compiler and added some opcodes to handle direct addressing
