+++ src/index.js
@@ -3,17 +3,21 @@ String.prototype.lpad = function (length) {
   return ("0000" + this.toString()).slice(length * -1);
 };
 
-var CPU = require('./CPU');
+////////////////////////////////////////////////
 
-var output = document.getElementById('output');
+var asmparser = require('./asmparser');
+
+var CPU = require('./CPU');
 var cpu = new CPU({
-  output: output
+  output: document.getElementById('output')
 });
+window.cpu = cpu;
 
 var Keyboard = require('./Devices/Keyboard');
 // Install keyboard with buffer on 0xF5E1 and 256 bytes of buffer
 cpu.installDevice(Keyboard, 0xF5E0, 256);
 
+// Install sys interrupts
 var sysInts = require('./SysInts');
 cpu.assignInterrupt(0x1, sysInts)
 
@@ -47,78 +51,87 @@ cpu.log = function () {
   }
 };
 
+// Instantiate memviewer
 var MemViewer = require('./MemViewer');
 var mv = new MemViewer(document.getElementById('memMap'), cpu.memory._raw, 0);
-cpu.clock.on('tick', function () {
-  mv.render();
+mv.clock.on('tick', function () {
   document.getElementById('memMap-from').innerHTML = mv.offset.toString(16).lpad(4);
   document.getElementById('memMap-to').innerHTML = (mv.offset + mv.viewportLength).toString(16).lpad(4);
 });
+mv.start();
+window.mv = mv;
 
-var program = new Uint8Array([
-  
-  // Type in screen
-  
-  // Screen mem in X / Cursor pos
-  42, 4, 0xE0,  // MOV XL, 0xE0
-  42, 5, 0xF6,  // MOV XH, 0xF6
-  
-  // Screen limit in Y
-  // off: 6
-  42, 6, 0xF7,  // MOV YL, 0xF7
-  42, 7, 0x30,  // MOV YH, 0x30
-  
-  // Add prompt
-  // off: 12
-  45, 20, 0xC5, // MOV [X], 0xC5
-  
-  // Check keystrokes
-  // off: 15
-  42, 0, 0,     // MOV A, 0x0
-  81, 0,        // INT 0x0; A=0
-  76, 3, 0,     // CMP D, 0 ; No new keys
-  57, 15,        // JE 15; Jump to loop
-  
-  // off: 25
-  42, 0, 3,     // MOV A, 0x3 ; Copy keyboard buffer[B] to X, ORed with C
-  42, 1, 0,     // MOV B, 0
-  42, 2, 0xC0,  // MOV C, 0xC0; OR with white
-  81, 0,        // INT 0x0; A=3
-  
-  // off: 36
-  42, 0, 4,     // MOV A, 0x4; Force shift
-  81, 0,        // INT 0x0; A=4
-  42, 3, 0,     // MOV D, 0x0
-  17, 20,       // INC X
-  
-  // off: 46
-  73, 20, 22,   // CMP X, Y
-  57, 53,       // JE 53
-  54, 72,       // JMP 72
-  
-  // off: 53
-  47, 2,          // PUSH C
-  42, 2, 80,      // MOV C, 80; Counter to 80
-  42, 0, 5,       // MOV A, 0x5; Force shift left
-  81, 0,          // INT 0x0; A=5
-  18, 2,          // DEC C
-  73, 2, 0,       // CMP C, 0
-  60, 59,         // JNE 59
-  51, 2,          // POP C
-  
-  // off: 72
-  45, 20, 0xC5, // MOV [X], 0xC5
-  54, 15         // JMP 15; Return to main loop
+// Bind some UI buttons
+document.getElementById('ctrl-speed-up').addEventListener('click', function () {
+  cpu.clock.speed += 10;
+  doLog('Set speed to', cpu.clock.speed);
+});
+document.getElementById('ctrl-speed-down').addEventListener('click', function () {
+  cpu.clock.speed -= 10;
+  doLog('Set speed to', cpu.clock.speed);
+});
+document.getElementById('ctrl-speed-set').addEventListener('click', function () {
+  cpu.clock.speed = parseInt(prompt('Set new clock speed:', '1000'), 10);
+  doLog('Set speed to', cpu.clock.speed);
+});
+document.getElementById('ctrl-mv-up').addEventListener('click', function () {
+  try {
+    var off = mv.offset, newOffset = off + mv.viewportLength;
+    mv.offset = newOffset;
+    doLog('Set memviewer to', mv.offset.toString(16).lpad(4));
+  } catch (e) {
+    doLog('Memory limit reached in memviewer');
+  }
+});
+document.getElementById('ctrl-mv-down').addEventListener('click', function () {
+  try {
+    var off = mv.offset, newOffset = off - mv.viewportLength;
+    mv.offset = newOffset;
+    doLog('Set memviewer to', mv.offset.toString(16).lpad(4));
+  } catch (e) {
+    doLog('Memory limit reached in memviewer');
+  }
+});
+document.getElementById('ctrl-mv-set').addEventListener('click', function () {
+  var newOffset = parseInt(prompt('Set new memview offset (hex): 0x', 'F5E0'), 16);
+  if (isNaN(newOffset)) {
+    doLog('Error! Offset', newOffset, 'is not a number');
+    return;
+  }
+  mv.offset = newOffset;
+  doLog('Set offset to', mv.offset);
+});
 
-]);
-
-window.cpu = cpu;
-cpu.loadProgram(program);
-
+// Bind hotkeys
 window.addEventListener('keydown', function (e) {
   //console.log(e.which);
+  // Ctrl C Compile
+  if (e.ctrlKey && e.which === 67) {
+    e.preventDefault();
+    e.stopPropagation();
+    e.stopImmediatePropagation();
+    cpu.halt();
+    try {
+      cpu.reset();
+      var r = asmparser(document.getElementById('code-txt').value);
+      cpu.loadProgram(Uint8Array.from(r.bytecode));
+    } catch (e) {
+      console.error(e);
+      return;
+    }
+  }
+  // Ctrl R to reset
+  if (e.ctrlKey && e.which === 82) {
+    e.preventDefault();
+    e.stopPropagation();
+    e.stopImmediatePropagation();
+    cpu.reset();
+  }
   // Ctrl P to Pause/resume
   if (e.ctrlKey && e.which === 80) {
+    e.preventDefault();
+    e.stopPropagation();
+    e.stopImmediatePropagation();
     if (cpu.clock.timer !== null) {
       cpu.halt();
     } else {
@@ -129,64 +142,11 @@ window.addEventListener('keydown', function (e) {
   }
   // Ctrl D to debug
   if (e.ctrlKey && e.which === 68) {
+    e.preventDefault();
+    e.stopPropagation();
+    e.stopImmediatePropagation();
     cpu.toggleDebug();
     doLog('Set debug to:', cpu.debug);
-    e.preventDefault();
-    return false;
-  }
-  if (e.shiftKey && e.which === 40) {
-    cpu.clock.speed += 10;
-    doLog('Set speed to', cpu.clock.speed);
-    e.preventDefault();
-    return false;
-  }
-  if (e.shiftKey && e.which === 38) {
-    if (cpu.clock.speed > 0) {
-      cpu.clock.speed -= 10;
-      doLog('Set speed to', cpu.clock.speed);
-      e.preventDefault();
-    }
-    return false;
-  }
-  if (e.shiftKey && e.which === 39) {
-    cpu.clock.speed = parseInt(prompt('Set new clock speed:', '1000'), 10);
-    doLog('Set speed to', cpu.clock.speed);
-    e.preventDefault();
-    return false;
-  }
-  if (e.altKey && e.which === 40) {
-    try {
-      var off = mv.offset, newOffset = off + mv.viewportLength;
-      mv.offset = newOffset;
-      doLog('Set memviewer to', mv.offset.toString(16).lpad(4));
-    } catch (e) {
-      doLog('Memory limit reached in memviewer');
-    }
-    e.preventDefault();
-    return false;
-  }
-  if (e.altKey && e.which === 38) {
-    try {
-      var off = mv.offset, newOffset = off - mv.viewportLength;
-      mv.offset = newOffset;
-      doLog('Set memviewer to', mv.offset.toString(16).lpad(4));
-    } catch (e) {
-      doLog('Memory limit reached in memviewer');
-    }
-    e.preventDefault();
-    return false;
-  }
-  if (e.altKey && e.which === 39) {
-    var newOffset = parseInt(prompt('Set new memview offset (hex): 0x', 'F5E0'), 16);
-    if (isNaN(newOffset)) {
-      doLog('Error! Offset', newOffset, 'is not a number');
-      return;
-    }
-    mv.offset = newOffset;
-    doLog('Set offset to', mv.offset);
-    e.preventDefault();
     return false;
   }
 });
-
-cpu.run();
